Improve destroy_later implementation
* Fix CHANGELOG entry

* Avoid using abreviations in argument names

* Refactor the job to use `find_by`

* Improve documentation

* Check if both jobs are configured when calling destroy_later at class

* Pass keyword arguments to the job to avoid warnings on Ruby 2.7

* Remove duplicated method

* Fix typos and avoid adding new tables

* Move active record queues configuration to Active Record

This config is related to jobs but it doesn't load anything on Active
Job, so there is no need to move the config to the Active Job railtie.

* Allow to configure the destroy later jobs per model

We are still going to configure this globaly to the framework but if
some user wants to use a different job in a model the system will work.
This also make easier to isolate the Active Job tests from the Active
Record jobs.

* Change the tests description to talk about the framework features

The tests were talking about the model features, not which features in
the framework they were asserting.

* Add documentation about how to configure the destroy jobs
