+++ activejob/lib/active_job/railtie.rb
@@ -38,7 +38,6 @@ module ActiveJob
       ActiveSupport.on_load(:active_record) do
         self.destroy_association_later_job = ActiveRecord::DestroyAssociationLaterJob
         self.destroy_later_job = ActiveRecord::DestroyJob
-        self.queues = ActiveSupport::InheritableOptions.new(destroy: :active_record_destroy)
       end
     end
