+++ activerecord/lib/active_record/associations/association.rb
@@ -342,10 +342,10 @@ module ActiveRecord
             reflection.source_reflection.active_record.default_scopes.any?
         end
 
-
         def enqueue_destroy_association(options)
-          ActiveRecord::Base.destroy_association_later_job.
-            perform_later(options) if ActiveRecord::Base.destroy_association_later_job
+          klass = owner.class
+          klass.destroy_association_later_job.
+            perform_later(**options) if klass.destroy_association_later_job
         end
     end
   end
