+++ activerecord/lib/active_record/associations/belongs_to_association.rb
@@ -13,14 +13,16 @@ module ActiveRecord
           raise ActiveRecord::Rollback unless target.destroyed?
         when :destroy_later
           id = owner.send(reflection.foreign_key.to_sym)
-          primary_key_column = reflection.
-              active_record_primary_key.to_sym
-          enqueue_destroy_association(owner_model_name: owner.class.to_s,
-                          owner_id: owner.id,
-                          assoc_class: reflection.klass.to_s,
-                          assoc_ids: [id],
-                          assoc_primary_key_column: primary_key_column,
-                         owner_ensuring_destroy_method: options.fetch(:owner_ensuring_destroy, nil))
+          primary_key_column = reflection.active_record_primary_key.to_sym
+
+          enqueue_destroy_association(
+            owner_model_name: owner.class.to_s,
+            owner_id: owner.id,
+            association_class: reflection.klass.to_s,
+            association_ids: [id],
+            association_primary_key_column: primary_key_column,
+            owner_ensuring_destroy_method: options.fetch(:owner_ensuring_destroy, nil)
+          )
         else
           target.send(options[:dependent])
         end
