+++ activerecord/lib/active_record/associations/has_one_association.rb
@@ -35,13 +35,15 @@ module ActiveRecord
           when :destroy_later
             primary_key_column = target.class.primary_key.to_sym
             id = target.send(primary_key_column)
-            enqueue_destroy_association(owner_model_name: owner.class.to_s,
-                            owner_id: owner.id,
-                            assoc_class: reflection.klass.to_s,
-                            assoc_ids: [id],
-                            assoc_primary_key_column: primary_key_column,
-                           owner_ensuring_destroy_method: options.fetch(:owner_ensuring_destroy, nil))
 
+            enqueue_destroy_association(
+              owner_model_name: owner.class.to_s,
+              owner_id: owner.id,
+              association_class: reflection.klass.to_s,
+              association_ids: [id],
+              association_primary_key_column: primary_key_column,
+              owner_ensuring_destroy_method: options.fetch(:owner_ensuring_destroy, nil)
+            )
           when :nullify
             target.update_columns(nullified_owner_attributes) if target.persisted?
           end
