+++ activerecord/test/activejob/destroy_later_job_not_present_test.rb
@@ -3,14 +3,13 @@
 require "cases/helper"
 
 class UnusedDestroyLater < ActiveRecord::Base
+  self.destroy_association_later_job = nil
+  self.destroy_later_job = nil
 end
 
 class UnusedBelongsTo < ActiveRecord::Base
+  self.destroy_association_later_job = nil
 end
-
-class UnusedHasMany < ActiveRecord::Base
-end
-
 
 class ActiveJobNotPresentTest < ActiveRecord::TestCase
   test "destroy later raises exception when activejob is not present" do
