+++ activerecord/test/models/category.rb
@@ -16,7 +16,6 @@ class Category < ActiveRecord::Base
                           -> { where title: "Yet Another Testing Title" },
                           class_name: "Post"
 
-
   has_and_belongs_to_many :popular_grouped_posts, -> { group("posts.type").having("sum(comments.post_id) > 2").includes(:comments) }, class_name: "Post"
   has_and_belongs_to_many :posts_grouped_by_title, -> { group("title").select("title") }, class_name: "Post"
