+++ activerecord/test/models/essay_destroy_later.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 class EssayDestroyLater < ActiveRecord::Base
-  belongs_to :book_destroy_later, dependent: :destroy_later
+  self.table_name = "essays"
+  belongs_to :book, dependent: :destroy_later, class_name: "BookDestroyLater"
 end
